fire_detection/requirements.txt
opencv-python 
numpy 
pyyaml 
twilio